containers: Remove Docker Apps support
